+++ lib/shared_preferences_explorer.dart
@@ -1,1 +1,2 @@
 export 'src/shared_preferences_explorer.dart';
+export 'src/shared_preferences_explorer_screen.dart';