+++ lib/src/floating_snack_bar.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-SnackBar floatingSnackBar({required String message}) {
-  return SnackBar(
-    content: Text(message),
-    behavior: SnackBarBehavior.floating,
-    showCloseIcon: true,
-  );
-}