+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0
+
+* Removes edit and delete functionality.
+* Adds `SharedPreferencesExplorerScreen`.
+* Supports `SharedPreferencesAsync` and `SharedPreferencesWithCache`.
+* Makes the anchor button draggable.
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+
 ## 0.0.2
 
 * Update README